pyproject.toml
[project]
name = "borjigin-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "chromadb>=1.3.4",
    "langchain>=1.0.4",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.3",
    "langchain-openai>=1.0.2",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.2",
    "langgraph-checkpoint>=3.0.1",
    "newspaper3k>=0.2.8",
    "openai>=2.7.1",
    "pandas>=2.3.3",
    "playwright>=1.55.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tavily-python>=0.7.12",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.14.4",
]
